Add ClientXSI
Create ClientXSI and ClientSync SHMs from the server where
the client receives the SHM key and maps it.

Break out work to be done from boost::asio::steady_timer.
This is more canonical way of scheduling periodic work because:

Timers automatically cancel cleanly when io.stop() is called,
and you won’t recurse infinitely or post after stop.

With a recursive call we would have to pass a
std::atomic<bool>& stopRequested and check it.